refactor(tools): give the child-execution stream-id formula an owner and split ExecutionsTool
Three related tool-layer cleanups.

(a) `src/tools/ExecutionsTool.ts` mixed action dispatch, store reads, and
five self-contained projections in one 1532-line file. The projections move
verbatim to `src/tools/executions/*` (joining `executionKvFiles.ts` and
`runGeneratedFiles.ts`): the resource-path catalog, the workflow-snapshot
view, the background-command output projection, the Zod action union, the
config field filter, and the turn-attribution note. No barrel; the router
imports each defining file. Router: 1532 -> 978 lines.

(b) Two delegation launch sites each derived the child's stream id and
assembled the same `registerOwnedExecution` options, guarded only by a
comment asking the copies to agree. `registerChildExecution` in
`detachedChildRun.ts` now mints it once. `WorkflowScriptTool` keeps its own
registration: it registers a fabricated container record under a
`multiAgentWorkflow` identity with a `getChildStreamId` prefix id, so it is
not the same ceremony.

(c) `.nullish().transform((v) => v ?? DEFAULT)` appeared verbatim at 11 tool
input fields. `nullishWithDefault` in `src/tools/core/inputSchema.ts` states
once why the wrapping is `.nullish()` and not `.prefault()` — the latter
substitutes only for `undefined`, and OpenAI-compatible providers send an
explicit `null` for an omitted optional field. Emitted JSON Schema and parse
results are identical for absent, null, and undefined input.
